Update README.md
Why tell ppl to not use if(statment) then, insted of if statment then, the code looks more clearer and even Mark admited that it does not slow anything down. And since most of the scripts are made with if(statment) then.. both me and printer do it like that, why not continue it?