Hot Cuisine Quest! Done! Dish 5-14 & Fix!
Added Jean Pierre on the map in this pos: 33067, 32531, 6.
Added Aid/Uid to door and chest: 33066, 32531, 5 and added the cookbook
item into the chest!
I have no idea where this startup error came from, but it said missing
Orc Berseker.xml npc, so i added it back.

Todo:
To make the dishes to work.

Former-commit-id: 114f709009d0448d1ec2a62e7652d961aa56c493