Rework The Oracle
Now oracle is working, but note: The npc was changed in the Update 8.2
when the Island of Destiny was introduced.

Thank you also for the firstitems.lua, i was able to merge the code into
The Oracle.

The rookgaard is under progress, when we reach to that point we shall
updated The Oracle Again.