Rework The Oracle
Now oracle is working, but note: The npc was changed in the Update 8.2
when the Island of Destiny was introduced.

Thank you also for the firstitems.lua, i was able to merge the code into
The Oracle.

The rookgaard is under progress, when we reach to that point we shall
updated The Oracle Again.


Former-commit-id: 4eafdc0ae6692aa85e319a1673f63845b0428594 [formerly 9b11a3486d38da1e891d1ee54b1e3b12808ac4e0]
Former-commit-id: edb2db675810df69f7a671ecb88ecd4282995775