Login Protection & Anti-kick feature
Anti-kick feature
In the case of disconnection from server, players now have a greater
chance of survival.

Characters log off after a shorter period of time depending on vocation:
a knight (or characters of no vocation) will log off after 60 seconds, a
paladin 50 and a druid/sorcerer 40. Before, it was 60 seconds for all
vocations.
Characters are now ignored by creatures for 30 seconds upon the time the
server recognizes the client has lost connection. This leaves less time
a character is vulnerable: 30 seconds (knight), 20 seconds (paladin) or
10 seconds (druid/sorcerer).